Version 1.11.2, changes below:
(Python) adjust tlpat to terminate folding on [A..Z]. Note that this
results in folding breaking with multi-line docstrings, which is what
the tlpat logic tried to fix in the first place.

Added some error handling to FoldBegin to handle gracefully cases when
the foldmethod is "indent" and when the fold_function is undefined.